requirements.txt
requests
BeautifulSoup4
python-telegram-bot
vk_api
shazamio
pyTelegramBotAPI
aiogram
pycryptodome
